Try to do full-screen mode correctly
Unfortunately this causes worse visual glitches than the existing code
when the game changes screen resolution during full-screen mode. We'll
need to change emscripten so that it calls a method on the Module
whenever the screen resolution changes, much as it does for things like
printing. This will allow us to do everything correctly.